HeaderLogo 컴포넌트 분리, 쓸모없는 파일 삭제
