+++ client/.eslintrc.cjs
@@ -7,6 +7,7 @@ module.exports= {
   rules: {
     "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
     "react/prop-types":"off",
+    "react/react-in-jsx-scope":"off",
   },
   "extends": ['airbnb'],
 };
