+++ client/src/common/HorizontalContainer.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const HorizontalContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
+
+export default HorizontalContainer;
