+++ client/src/common/VerticalContainer.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const VerticalContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
+
+export default VerticalContainer;
